Move files/folders around for better readability.
Change the feature store to instantiate as a singleton, and the clients to get an instance rather than instantiate their own.
Get the feature path from static information at startup time.
Change the initial type expected in the Gherkin parsing to be None rather than Unknown, and filtered tabs out (in a very quick fashion!) from the input string.